[fix]quant

+++ models/quant.py
@@ -13,14 +13,23 @@ class Quant(models.Model):
     block_id = fields.Many2one('product.block', '荒料编号', related='product_id.block_id', store=True, readonly=True)
     location_id = fields.Many2one('stock.location', '库存位置', auto_join=True, ondelete='restrict', readonly=True,
                                   required=True)
-    pcs = fields.Integer('件数')
+    pcs = fields.Integer('件数', required=True, default=0)
     part = fields.Integer('夹数', compute='_compute_qty', store=True)
-    qty = fields.Float('数量', readonly=True, compute='_compute_qty', store=True)
+    qty = fields.Float('数量', compute='_compute_qty', store=True)
     uom = fields.Many2one('product.uom', '单位', related='product_id.uom', store=True, readonly=True)
     slab_ids = fields.One2many('product.slab', 'quant_id', string='板材')
+
+    reserved_pcs = fields.Integer('预留件数')
+    reserved_qty = fields.Float('预留数量', compute='_compute_qty', store=True)
+    reserved_slab_ids = fields.Float('预留板材')
+
     in_date = fields.Datetime('入库日期', readonly=True)
 
     _sql_constraints = [('unique_product_location', 'unique (product_id,location_id)', '该位置已有该产品,不能创建同位置同产品的库存框!')]
+
+    def _check_product_type(self, product_id, slab_ids=None):
+        if product_id.type == 'slab' and slab_ids is None:
+            raise exceptions.ValidationError('产品[ {} ]的形态为:板材,没有附带码单.不能对库存操作!')
 
     @api.depends('product_id.type', 'slab_ids')
     def _compute_qty(self):
@@ -29,8 +38,9 @@ class Quant(models.Model):
                 record.qty = sum(record.mapped('slab_ids.m2'))
                 record.pcs = len(record.slab_ids)
                 record.part = len(set(record.mapped('slab_ids.part_num')))
+                record.reserved_qty = sum(record.mapped('reserved_slab_ids.m2'))
             elif record.product_id.type == 'pbom':
-                record.qty = record.pcs * record.qty
+                record.qty = record.pcs * record.product_id.single_qty
                 record.part = None
             elif record.product_id.type == 'block':
                 record.qty = record.product_id.single_qty
@@ -47,21 +57,77 @@ class Quant(models.Model):
         return res
 
     @api.model
+    def _update_reserved(self, product_id, location_id, pcs, slab_ids, strict=False):
+        """
+        :param product_id:
+        :param location_id:
+        :param pcs:
+        :param slab_ids:
+        :param strict:
+        :return: tuple(quant(recordset), pcs(int), slab_ids(list))
+        """
+        self._check_product_type(product_id, slab_ids)
+        quants = self._get(product_id, location_id, strict=strict)
+        # available = self._get_available(product_id, location_id)
+        reserved_quants = []
+        if slab_ids:
+            quantity = len(slab_ids.mapped('id')) * -1 if pcs < 0 else len(slab_ids.mapped('id'))
+        for quant in quants:
+            if quantity > 0:
+                max_reserved_pcs = quant.pcs - quant.reserved_pcs
+                if max_reserved_pcs <= 0:
+                    continue
+                if slab_ids:
+                    quant.reserved_slab_ids = slab_ids & quant.slab_ids
+                    slab_ids -= quant.reserved_slab_ids
+                max_reserved_pcs = min(max_reserved_pcs, quantity)
+                quant.reserved_pcs -= max_reserved_pcs
+                reserved_quants.append((quant, max_reserved_pcs, quant.mapped('reserved_slab_ids.id')))
+                quantity -= max_reserved_pcs
+            else:
+                max_reserved_pcs = min(quant.reserved_pcs, abs(quantity))
+                if slab_ids:
+                    reserved_slab_ids = quant.reserved_slab_ids
+                    quant.reserved_slab_ids -= slab_ids
+                    slab_ids -= quant.reserved_slab_ids
+                quant.reserved_pcs -= max_reserved_pcs
+                reserved_quants.append((quant, quantity, reserved_slab_ids.mapped('id')))
+                quantity += max_reserved_pcs
+            if quantity == 0:
+                break
+        return reserved_quants
+
+    @api.model
+    def _get_available(self, product_id, location_id):
+        """
+        :param product_id:(record)
+        :param location_id: location(record)
+        :return: tuple(pcs(int), slabs(list))
+        """
+        quants = self.get(product_id, location_id)
+        slabs = [quant.mapped('slab_ids.id') for quant in quants]
+        pcs = sum(quant.pcs for quant in quants)
+        return pcs, slabs
+
+    @api.model
     def _update_available(self, product_id, location_id, pcs, slab_ids=None):
+        self._check_product_type(product_id, slab_ids)
         self = self.sudo()
         if product_id.type == 'slab':
             if not slab_ids:
                 raise exceptions.ValidationError('更新库存位置的板材产品没有附带码单!')
         quants = self._get(product_id, location_id, strict=True)
+        balance = 0
         for quant in quants:
-            if (quant.pcs + pcs) < 0:
-                raise exceptions.ValidationError('移出的库存数量不能大于库存数量!')
+            # if (quant.pcs + pcs) < 0:
+            #     raise exceptions.ValidationError('移出的库存数量不能大于库存数量!')
+            balance += quant.pcs + pcs
             vals = {'pcs': quant.pcs + pcs}
             if slab_ids:
                 slabs = quants.slab_ids - slab_ids
                 vals['slab_ids'] = [(6, 0, slabs.mapped('id'))]
             quant.write(vals)
-            if quant.pcs == 0:
+            if quant.pcs == 0 and quant.reserved_pcs == 0:
                 quants.unlink()
             break
         else:
